+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][keepachangelog],
 and this project adheres to [Semantic Versioning][semver].
 
+## [1.1.3]
+
+- Dependency updates: need pygls 1.0.1, drop support for Python 3.6
+- Changed settings loading method.
+
 ## [1.1.2]
 
 - Fixed python version checking
@@ -65,6 +70,7 @@ and this project adheres to [Semantic Versioning][semver].
 [keepachangelog]: https://keepachangelog.com/en/1.0.0/
 [semver]: https://semver.org/spec/v2.0.0.html
 
+[1.1.3]: https://github.com/HankBO/stata-language-server/compare/v1.1.2...v1.1.3
 [1.1.2]: https://github.com/HankBO/stata-language-server/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/HankBO/stata-language-server/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/HankBO/stata-language-server/compare/v1.0.3...v1.1.0